+++ sync-for-commerce/RELEASES.md
@@ -529,3 +529,13 @@ Based on:
 - [python v0.31.0] sync-for-commerce
 ### Releases
 - [PyPI v0.31.0] https://pypi.org/project/codat-sync-for-commerce/0.31.0 - sync-for-commerce
+
+## 2023-09-08 00:03:52
+### Changes
+Based on:
+- OpenAPI Doc 1.1 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Commerce.yaml
+- Speakeasy CLI 1.82.0 (2.107.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v1.1.0] sync-for-commerce
+### Releases
+- [PyPI v1.1.0] https://pypi.org/project/codat-sync-for-commerce/1.1.0 - sync-for-commerce